Split tracking test conftest
